Suggestion: Automatic Wallet Unlock
Comments
I am proposing this understanding it is very controversial to embed a password in a plain text file.

In the testnet environment - there is no real risk
In the mainnet environment - there is a risk. It would be up to the system manager to decide if to include this or not.